B-1 refactoring containers.
