+++ src/component/Button.js
@@ -5,7 +5,7 @@
 
 import React, { Component } from 'react';
 
-class Button extends Component {
+export default class Button extends Component {
   constructor(props){
     super(props);
     this.state={
@@ -20,5 +20,3 @@ class Button extends Component {
     );
   }
 }
-
-export default Button;
