+++ src/component/Logo.js
@@ -5,7 +5,7 @@
 
 import React, { Component } from 'react';
 
-class Logo extends Component {
+export default class Logo extends Component {
   render(){
     return(
       <div>
@@ -39,5 +39,3 @@ class Logo extends Component {
     );
   }
 }
-
-export default Logo;
