+++ src/container/login-form-container.js
@@ -3,16 +3,20 @@
  * Copyright 2017 Prottonne
  */
 
-import { login } from '../action/auth-action';
+import { login } from '../action/user-action';
+import LoginForm from '../component/LoginForm';
+import { connect } from 'react-redux';
 
- export const mapStateToProps = (store) => {
+ const mapStateToProps = (store) => {
    return {
      user:store.user
    };
  }
 
- export const matchDispatchToProps = (dispatch) => {
+ const matchDispatchToProps = (dispatch) => {
    return   {
      login: (user) => dispatch(login(user))
    };
  }
+
+ export default connect(mapStateToProps, matchDispatchToProps )(LoginForm);
