+++ src/container/password-input-container.js
@@ -4,15 +4,19 @@
  */
 
  import { setUserPassword } from '../action/user-action';
+ import PasswordInput from '../component/PasswordInput';
+ import { connect } from 'react-redux';
 
-  export const mapStateToProps = (store) => {
+  const mapStateToProps = (store) => {
     return {
       user:store.user
     };
   }
 
-  export const matchDispatchToProps = (dispatch) => {
+  const matchDispatchToProps = (dispatch) => {
     return   {
       setUserPassword: (password) => dispatch(setUserPassword(password))
     };
   }
+
+export default connect(mapStateToProps, matchDispatchToProps)(PasswordInput);
