+++ src/route/HomePage.js
@@ -4,11 +4,9 @@
  */
 
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { mapStateToProps } from '../container/home-page-container';
 import { Link } from 'react-router-dom';
 
-class HomePage extends Component {
+export default class HomePage extends Component {
   render(){
 
     if(this.props.user.isAuth){
@@ -30,5 +28,3 @@ class HomePage extends Component {
     }
   }
 }
-
-export default connect(mapStateToProps)(HomePage);
